+++ lib/recode/formatter_plugin.ex
@@ -47,7 +47,6 @@ defmodule Recode.FormatterPlugin do
   def features(opts) do
     # This callback will be applied for every file the Elixir formater wants to
     # format. All calls comming from one Process.
-    # Here it is a little misappropriated as `&init/1`.
     _ref = init(opts)
 
     [extensions: [".ex", ".exs"]]
@@ -55,6 +54,8 @@ defmodule Recode.FormatterPlugin do
 
   @impl true
   def format(content, formatter_opts) do
+    file = formatter_opts |> Keyword.get(:file, "source.ex") |> Path.relative_to_cwd()
+
     formatter_opts =
       Keyword.update(formatter_opts, :plugins, [], fn plugins ->
         Enum.reject(plugins, fn plugin -> plugin == Recode.FormatterPlugin end)
@@ -62,7 +63,7 @@ defmodule Recode.FormatterPlugin do
 
     config = Keyword.put(config(), :dot_formatter_opts, formatter_opts)
 
-    Recode.Runner.run(content, config)
+    Recode.Runner.run(content, config, file)
   end
 
   defp config do